Initial inclusion of deliberately butchered LibreSSL
We're including this in Solus only with the intent of having LSI perform
intercepts on LibreSSL to ensure we get these games to use updated versions
of LibreSSL, i.e. for security.

We don't actually use LibreSSL ourselves in Solus so I've opted to make sure
nobody can build against these packages, as well as renaming the libraries
on disk so that nothing would ever intentionally look for these libraries
without being told to first by lsi-redirect.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>